scripts/requirements.txt
# TF.js CNN Regressor Training Dependencies
# Python 3.9+

# Core ML framework
tensorflow==2.15.0
tensorflowjs==4.17.0

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Data handling
numpy==1.26.4
pandas==2.2.0

# Supabase client
supabase==2.3.4

# AWS S3/R2 client
boto3==1.34.51

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
